check user access

+++ src/controllers/devicesController.js
@@ -4,7 +4,6 @@ const { validationResult } = require('express-validator');
 const { errorMsgTrans } = require('../utils/transform');
 const Device = require('../models').device;
 const { v4: uuidv4 } = require('uuid');
-const { token } = require('morgan');
 
 exports.getDevices = async (req, res) => {
     const { id: user_id } = req.user;
@@ -24,14 +23,22 @@ exports.getDevices = async (req, res) => {
 
 exports.getDevice = async (req, res) => {
     const { id } = req.params;
+    const { id: user_id } = req.user;
 
     //get device detail by id
-    const deviceDetail = await Device.findByPk(id);
-
+    const deviceDetail = await Device.findAll({ where: { id } });
     if (!deviceDetail)
         throw new AppError('The id is not related to any devices', 404);
 
-    res.json({ code: 200, status: 'success', data: deviceDetail });
+    // check if user has access
+    if (!deviceDetail.user_id === user_id)
+        throw new AppError('Access forbidden!', 403);
+
+    res.json({
+        code: 200,
+        status: 'success',
+        data: deviceDetail,
+    });
 };
 
 exports.addDevice = async (req, res) => {
@@ -76,6 +83,7 @@ exports.addDevice = async (req, res) => {
 
 exports.editDevice = async (req, res) => {
     const { id } = req.params;
+    const { id: user_id } = req.user;
     const bodyData = req.body;
 
     // check body payload
@@ -93,9 +101,13 @@ exports.editDevice = async (req, res) => {
     }
 
     //find device to update
-    const device = await Device.findByPk(id);
+    const device = await Device.findAll({ where: { id } });
 
     if (!device) throw new AppError('Device not found!', 401);
+
+    // check if user has access
+    if (!device.user_id === user_id)
+        throw new AppError('Access forbidden!', 403);
 
     //update device
     const updateDevice = await device.update({
@@ -114,11 +126,15 @@ exports.editDevice = async (req, res) => {
 
 exports.deleteDevice = async (req, res) => {
     const { id } = req.params;
+    const { id: user_id } = req.user;
 
     //find device
-    const device = await Device.findByPk(id);
-
+    const device = await Device.findAll({ where: { id } });
     if (!device) throw new AppError('Device not found!', 404);
+
+    // check if user has access
+    if (!deviceDetail.user_id === user_id)
+        throw new AppError('Access forbidden!', 403);
 
     //delete device
     const deleteDevice = await device.destroy({
